TMDB: creating serie

+++ Application/Model/Timeline.cs
@@ -2,7 +2,7 @@
 {
     public class Timeline
     {
-        public short TimelineId { get; set; }
+        public byte TimelineId { get; set; }
         public string Name { get; set; }
         public virtual ICollection<Movie> Movies { get; set; }
         public virtual ICollection<Serie> Series { get; set; }
